Create jitpack.yml to set custom openjdk version
The #5 PR introduced a new compile configuration to include the module
information introduced in Java 9.
To include the module information the project has to be compiled with a
Java 9+ version and be recompiled excluding `module-info.java` while
targeting Java 8.

JitPack uses Java 8 as default version and therefore it fails building
it... This config should tell JitPack to use `openjdk9` as Java Version.

Refs:
- https://jitpack.io/docs/BUILDING/#java-version